requirements.txt
msgpack_python
tornado
mockito
opster >= 4.0